refactor: migrate avatars to typescript
